src/test/javascript/spec/app/entities/signature-field-parallel/signature-field-parallel.service.spec.ts
/* tslint:disable max-line-length */
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { HttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { take, map } from 'rxjs/operators';
import { SignatureFieldParallelService } from 'app/entities/signature-field-parallel/signature-field-parallel.service';
import { ISignatureFieldParallel, SignatureFieldParallel } from 'app/shared/model/signature-field-parallel.model';

describe('Service Tests', () => {
    describe('SignatureFieldParallel Service', () => {
        let injector: TestBed;
        let service: SignatureFieldParallelService;
        let httpMock: HttpTestingController;
        let elemDefault: ISignatureFieldParallel;
        beforeEach(() => {
            TestBed.configureTestingModule({
                imports: [HttpClientTestingModule]
            });
            injector = getTestBed();
            service = injector.get(SignatureFieldParallelService);
            httpMock = injector.get(HttpTestingController);

            elemDefault = new SignatureFieldParallel('ID', 0, 0, 0, 'image/png', 'AAAAAAA', 0);
        });

        describe('Service methods', async () => {
            it('should find an element', async () => {
                const returnedFromService = Object.assign({}, elemDefault);
                service
                    .find('123')
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: elemDefault }));

                const req = httpMock.expectOne({ method: 'GET' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should create a SignatureFieldParallel', async () => {
                const returnedFromService = Object.assign(
                    {
                        id: 'ID'
                    },
                    elemDefault
                );
                const expected = Object.assign({}, returnedFromService);
                service
                    .create(new SignatureFieldParallel(null))
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: expected }));
                const req = httpMock.expectOne({ method: 'POST' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should update a SignatureFieldParallel', async () => {
                const returnedFromService = Object.assign(
                    {
                        page: 1,
                        x: 1,
                        y: 1,
                        image: 'BBBBBB',
                        signerId: 1
                    },
                    elemDefault
                );

                const expected = Object.assign({}, returnedFromService);
                service
                    .update(expected)
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: expected }));
                const req = httpMock.expectOne({ method: 'PUT' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should return a list of SignatureFieldParallel', async () => {
                const returnedFromService = Object.assign(
                    {
                        page: 1,
                        x: 1,
                        y: 1,
                        image: 'BBBBBB',
                        signerId: 1
                    },
                    elemDefault
                );
                const expected = Object.assign({}, returnedFromService);
                service
                    .query(expected)
                    .pipe(
                        take(1),
                        map(resp => resp.body)
                    )
                    .subscribe(body => expect(body).toContainEqual(expected));
                const req = httpMock.expectOne({ method: 'GET' });
                req.flush(JSON.stringify([returnedFromService]));
                httpMock.verify();
            });

            it('should delete a SignatureFieldParallel', async () => {
                const rxPromise = service.delete('123').subscribe(resp => expect(resp.ok));

                const req = httpMock.expectOne({ method: 'DELETE' });
                req.flush({ status: 200 });
            });
        });

        afterEach(() => {
            httpMock.verify();
        });
    });
});
